// src/types/types.ts
export type PostsType = {
  id: number
  message: string
  name: string
  urlAvatar: string
}
export type ContactsType = {
  ithub: string | null
  vk: string | null
  facebook: string | null
  instagram: string | null
  twitter: string | null
  website: string | null
  youtube: string | null
  mainLink: string | null
}
export type PhotosType = {
  small: string | null
  large: string | null
}
export type ProfileType = {
  userId: number
  lookingForAJob: boolean
  lookingForAJobDescription: string
  fullName: string
  contacts: ContactsType
  photos: PhotosType
}
export type UsersType = {
  id: number
  name: string
  status: string
  photos: PhotosType
  followed: boolean
}